requirements.txt
streamlit
watchdog
platformdirs
pandas
st-theme
mpremote
pyserial
plotly
